refactor: rename db functions file name for evaluations
